% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_spat_pca.R
\name{auto_gau_basis_1d}
\alias{auto_gau_basis_1d}
\title{Gaussian kernel basis function generation (1-dim)}
\usage{
auto_gau_basis_1d(s_mat, num_center, scale = 1.5)
}
\arguments{
\item{s_mat}{location matrix}

\item{num_center}{number of center}

\item{scale}{sigma:delta = 1:scale, default is 1.5}
}
\value{
Gaussian kernel basis function
}
\description{
Gaussian kernel basis function generation (1-dim)
}
