% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_spat_pca.R
\name{B_gen}
\alias{B_gen}
\title{Generate B matrix}
\usage{
B_gen(s, phi_mat)
}
\arguments{
\item{s}{location coordinate matrix (generally a N by 2 matrix)}

\item{phi_mat}{basis function matrix}
}
\value{
B matrix in thesis p18
}
\description{
Generate B matrix
}
