% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{cv_result_all}
\alias{cv_result_all}
\title{Tuning results for sim_data}
\format{a list contains:
\describe{
\item{k_best}{the best K}
\item{tau_best}{the best tau}
\item{tau_list}{a list contains tuning process of tau in each K}
\item{k_process}{the negativet two log-likelihood in each K}
\item{best_list}{the detail of best tuning results}
}}
\usage{
cv_result_all
}
\description{
Tuning results for sim_data
}
\keyword{datasets}
