% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_spat_pca.R
\name{EM_func3}
\alias{EM_func3}
\title{Parameter estimation of Temporal Spatial PCA by EM-algorithm}
\usage{
EM_func3(s, y_mat, y_mat_new, sigma2_eps = 1, itermax = 30,
  tol = 0.001, tau, k)
}
\arguments{
\item{s}{location matrix}

\item{y_mat}{data matrix(training data)}

\item{y_mat_new}{data matrix(testing data)}

\item{sigma2_eps}{measurement error variance}

\item{itermax}{maximum iteration of EM-algorithm}

\item{tol}{stopping criterion of EM}

\item{tau}{roughness parameter}

\item{k}{dim of eigen-space}
}
\value{
a list of estimated parameters
}
\description{
Parameter estimation of Temporal Spatial PCA by EM-algorithm
}
