% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_spat_pca.R
\name{EM_func_model2}
\alias{EM_func_model2}
\title{Parameter estimation function for Model II}
\usage{
EM_func_model2(s, y_mat, y_mat_new, sigma2_eps = 1, itermax = 30,
  tol = 0.001, k, phi_mat)
}
\arguments{
\item{s}{location matrix}

\item{y_mat}{data matrix (training data)}

\item{y_mat_new}{data matrix (testing data)}

\item{sigma2_eps}{measurement error variance}

\item{itermax}{maximum number of iteration in EM-algorithm}

\item{tol}{stopping criterion of EM-algorithm}

\item{k}{dim of eigen-space K}

\item{phi_mat}{pre-specified basis function}
}
\value{
parameter estimates
}
\description{
Parameter estimation function for Model II
}
