% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_spat_pca.R
\name{lambda_mat_calc}
\alias{lambda_mat_calc}
\title{Calclulate lambda matrix by A & Ca}
\usage{
lambda_mat_calc(A, Ca)
}
\arguments{
\item{A}{state-transition matrix}

\item{Ca}{covariance matrix of random noise}
}
\value{
lambda matrix
}
\description{
Calclulate lambda matrix by A & Ca
}
