% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_spat_pca.R
\name{model1_func}
\alias{model1_func}
\title{Model I for simulation use}
\usage{
model1_func(s, s_new, y_mat, y_mat_new, xsi_mat, phi_mat, new_xsi_mat,
  phi_mat_new, lambda_mat, A)
}
\arguments{
\item{s}{location matrix for model fitting}

\item{s_new}{location matrix for spatial prediction}

\item{y_mat}{data matrix (training data)}

\item{y_mat_new}{data matrix (testing data)}

\item{xsi_mat}{state matrix (true value)}

\item{phi_mat}{phi matrix (true value)}

\item{new_xsi_mat}{new state matrix (true value)}

\item{phi_mat_new}{new phi matrix (true value)}

\item{lambda_mat}{state covariance matrix (true value)}

\item{A}{state transition matrix (true value)}
}
\value{
a list of parameter estimation result and performance results
}
\description{
Model I for simulation use
}
