% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{NCEP_U}
\alias{NCEP_U}
\title{Wind speed dataset obtained from Hsu, Chang, and Huang (2012)}
\format{a 480(time)*289(location) data matrix}
\usage{
NCEP_U
}
\description{
The data consists of half-yearly east-west component of the wind 
speed vector at 289 locations (17x17 regular grid) from November 1992 to February
1993 over latitudes $14^oS$ − $16^oN$ and longitudes $145^oE$ − $175^oE$ in the western Pacific ocean,
}
\keyword{datasets}
