% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{NCEPlatlon}
\alias{NCEPlatlon}
\title{Coordinate matrix of wind speed dataset}
\format{a 289(location)*2(coordinates) data matrix}
\usage{
NCEPlatlon
}
\description{
Coordinate matrix of wind speed dataset
}
\keyword{datasets}
