% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_spat_pca.R
\name{new_loc_phi}
\alias{new_loc_phi}
\title{Generate new Phi matrix}
\usage{
new_loc_phi(s_new, s_old, B)
}
\arguments{
\item{s_new}{locations for spatial prediction}

\item{s_old}{locations for model fitting}

\item{B}{B matrix in thesis p18}
}
\value{
generate phi(S0), in thesis p18
}
\description{
Generate new Phi matrix
}
