% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_spat_pca.R
\name{performance_func}
\alias{performance_func}
\title{Performance measure calculation from thesis p24}
\usage{
performance_func(y_new_true, y_spat, y_temp, lag0_cov_true, lag0_cov_est,
  lag1_cov_true, lag1_cov_est)
}
\arguments{
\item{y_new_true}{true signal matrix}

\item{y_spat}{spatial prediction matrix}

\item{y_temp}{temporal forecast matrix}

\item{lag0_cov_true}{true lag0 covariance matrix}

\item{lag0_cov_est}{estimated lag0 covariance matrix}

\item{lag1_cov_true}{true lag1 covariance matrix}

\item{lag1_cov_est}{estimated lag1 covariance matrix}
}
\value{
performance measures
}
\description{
Performance measure calculation from thesis p24
}
