% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_spat_pca.R
\name{pred_func}
\alias{pred_func}
\title{Prediction function of Temporal Spatial PCA}
\usage{
pred_func(new_phi, y_mat_new, EM_result)
}
\arguments{
\item{new_phi}{phi matrix for spatial prediction use}

\item{y_mat_new}{observation in finite locations in the future}

\item{EM_result}{a list created from EM_func3}
}
\value{
The results of Kalman filter (xi filter & prediction), 
spatial prediction & temporal forecast 
(both point & interval prediction)
}
\description{
Prediction function of Temporal Spatial PCA
}
\examples{
s = sim_data$location
s_new = sim_data$location_new
y_mat = sim_data$y_mat[1:100,]
y_mat_new = sim_data$y_mat[1:100,]
EM_result <- EM_func3(s=s, y_mat=y_mat, y_mat_new=y_mat_new, tau=8, k=2)
new_phi <- new_loc_phi(s_new=s_new, s_old = s, B = B_gen(s, EM_result$Phi))
pred_result = pred_func(new_phi=new_phi, y_mat_new=y_mat_new, EM_result=EM_result)
}
