% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulation data}
\format{a list contains:
\describe{
\item{y_mat}{simulation data matrix}
\item{xsi_mat}{random coefficient matrix}
\item{epsilon_mat}{measurement error matrix}
\item{location}{location matirx for fitting}
\item{phi_mat}{true phi matrix}
\item{location_new}{location matrix for prediction}
\item{phi_mat_new}{basis functions evaluated at location_new}
}}
\usage{
sim_data
}
\description{
generate
}
\keyword{datasets}
