% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_spat_pca.R
\name{temp_spat_cv_final3}
\alias{temp_spat_cv_final3}
\title{Tuning process of Temporal Spatial PCA}
\usage{
temp_spat_cv_final3(s, y_mat, sigma2_eps = 1, itermax = 30,
  tol = 0.001, tau, k)
}
\arguments{
\item{s}{location matrix}

\item{y_mat}{data matrix}

\item{sigma2_eps}{measurement error variance}

\item{itermax}{maximum iteration of EM-algorithm}

\item{tol}{stopping criterion of EM}

\item{tau}{a vector contains candidates of roughness parameter}

\item{k}{a vector contains candidates of eigen-space}
}
\value{
a list of tuning process
}
\description{
Tuning process of Temporal Spatial PCA
}
