% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_spat_pca.R
\name{temp_spat_data_em3}
\alias{temp_spat_data_em3}
\title{All in one function for data analysis}
\usage{
temp_spat_data_em3(s, s_new, y_mat, y_mat_new, sigma2_eps = 1,
  itermax = 30, tol = 0.001, tau, k)
}
\arguments{
\item{s}{location matrix}

\item{s_new}{location matrix for spatial prediction use}

\item{y_mat}{data matrix (training)}

\item{y_mat_new}{data matrix (testing)}

\item{sigma2_eps}{measurement error variance}

\item{itermax}{maximum iteration in EM-algorithm}

\item{tol}{stopping criterion}

\item{tau}{roughness parameter}

\item{k}{dim of eigen-space}
}
\value{
a list of parameter estimation and prediction results
}
\description{
All in one function for data analysis
}
