% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_spat_pca.R
\name{temp_spat_data_em_model2}
\alias{temp_spat_data_em_model2}
\title{Model II for data analysis use}
\usage{
temp_spat_data_em_model2(s, s_new, y_mat, y_mat_new, bas_mat, bas_mat_new,
  k, sigma2_eps = 1, itermax = 30, tol = 0.001)
}
\arguments{
\item{s}{location matrix for model fitting}

\item{s_new}{location matrix for spatial prediction}

\item{y_mat}{data matrix (training data)}

\item{y_mat_new}{data matrix (testing data)}

\item{bas_mat}{pre-specified basis functions evaluate at s}

\item{bas_mat_new}{pre-specified basis functions evaluate at s_new}

\item{k}{dim of eigen-space}

\item{sigma2_eps}{measurement error variance}

\item{itermax}{maximum iteration in EM-algorithm}

\item{tol}{stopping criterion}
}
\value{
a list of parameter estimation and prediction results
}
\description{
Model II for data analysis use
}
