% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_spat_pca.R
\name{V_mat_calc}
\alias{V_mat_calc}
\title{Calculate Ca matrix by lambda and A}
\usage{
V_mat_calc(A, lambda_mat)
}
\arguments{
\item{A}{state-transition matrix}

\item{lambda_mat}{state covariance matrix}
}
\value{
generate Ca, also known as Sigma_a, in thesis p10
}
\description{
Calculate Ca matrix by lambda and A
}
