\name{temp.spat.pca3-package}
\alias{temp.spat.pca3-package}
\alias{temp.spat.pca3}
\docType{package}
\title{
\packageTitle{temp.spat.pca3}
}
\description{
\packageDescription{temp.spat.pca3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{temp.spat.pca3}
\packageIndices{temp.spat.pca3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{temp.spat.pca3}

Maintainer: \packageMaintainer{temp.spat.pca3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
